src/scan-receipts/requirements.txt

annoy==1.17.3
app==0.0.1
fasttext-wheel==0.9.2
Flask==3.0.3
Flask-PyMongo==2.3.0
nltk==3.8.1
Pillow==10.0.1
pymongo==4.7.2
python-dotenv==1.0.1
requests==2.31.0
scipy==1.13.0
torch==2.3.0
transformers==4.40.1
